Bug 484411 - Remove XP_MACOSX and OS_TARGET=MacOSX. r=kaie
With support for macOS < 10.3 dropped, we don't need to care about
special cases for 10.1 anymore, which were the last places using
XP_MACOSX, which we can remove. And as per the previous change,
OS_TARGET=MacOSX is not used anymore. The value will now be Darwin
on all Darwin-based platforms. Also, as per the previous change,
we don't need to check for Carbon anymore.

Differential Revision: https://phabricator.services.mozilla.com/D86793